feat: support kb scoped tables migration
