+++ backend/app/rag/knowledge_base/index_store.py
@@ -2,9 +2,10 @@ from sqlalchemy import inspection
 from sqlmodel import Session
 
 from app.models import KnowledgeBase
+from app.models.chunk import get_kb_chunk_model
+from app.models.entity import get_kb_entity_model
 from app.rag.knowledge_base.config import get_kb_dspy_llm, get_kb_embed_model
-from app.rag.knowledge_base.db_model import get_kb_chunk_model, get_kb_entity_model, \
-    get_kb_relationship_model
+from app.models.relationship import get_kb_relationship_model
 from app.rag.knowledge_graph.graph_store import TiDBGraphStore
 from app.rag.knowledge_graph.graph_store.tidb_graph_editor import TiDBGraphEditor
 from app.rag.vector_store.tidb_vector_store import TiDBVectorStore
