+++ backend/app/tasks/build_index.py
@@ -10,10 +10,10 @@ from app.models import (
     DocIndexTaskStatus,
     KgIndexStatus,
 )
+from app.models.chunk import get_kb_chunk_model
 from app.models.knowledge_base import IndexMethod
 from app.rag.build_index import IndexService
 from app.rag.knowledge_base.config import get_kb_llm, get_kb_embed_model
-from app.rag.knowledge_base.db_model import get_kb_chunk_model
 from app.repositories import knowledge_base_repo
 from app.repositories.chunk import ChunkRepo
